Compiles in 64-bit configuration. Runs the intro movie (somewhat corrupted). Crashes on starting a game.

+++ src/Libraries/2D/Source/RSD/RSDUnpack.c
@@ -34,7 +34,7 @@ uchar *gr_rsd8_unpack(uchar *src, uchar *dest)
  	uchar		code,val;
  	short		count,count2;
  	ushort	longcode;
- 	ulong		longval, *longdest, *longsrc;
+ 	uint32_t		longval, *longdest, *longsrc;
  
  	do
  	 {
@@ -46,11 +46,11 @@ uchar *gr_rsd8_unpack(uchar *src, uchar *dest)
  		 	
  		 	if (count>=kMinLongLoop)		// if at least kMinLongLoop bytes, do long word stuff
  		 	 {
- 		 	 	longval = val + (((ulong) val)<<8);
+ 		 	 	longval = val + (((uint32_t) val)<<8);
  		 	 	longval += longval<<16;
  		 	 	count2 = count>>2;
  		 	 	count &= 3;
- 		 	 	longdest = (ulong *) dest;
+ 		 	 	longdest = (uint32_t *) dest;
  		 	 	
  		 	 	while (count2--)
  		 	 	 	*(longdest++) = longval;
@@ -68,8 +68,8 @@ uchar *gr_rsd8_unpack(uchar *src, uchar *dest)
  		 	 {
  		 	 	count2 = count>>2;
  		 	 	count &= 3;
- 		 	 	longdest = (ulong *) dest;
- 		 	 	longsrc = (ulong *) src;
+ 		 	 	longdest = (uint32_t *) dest;
+ 		 	 	longsrc = (uint32_t *) src;
  		 	 	
  		 	 	while (count2--)
  		 	 	 	*(longdest++) = *(longsrc++);
@@ -90,7 +90,7 @@ uchar *gr_rsd8_unpack(uchar *src, uchar *dest)
  		 	 {
  		 	 	count2 = count>>2;
  		 	 	count &= 3;
- 		 	 	longdest = (ulong *) dest;
+ 		 	 	longdest = (uint32_t *) dest;
  		 	 	
  		 	 	while (count2--)
  		 	 	 	*(longdest++) = longval;
@@ -116,7 +116,7 @@ uchar *gr_rsd8_unpack(uchar *src, uchar *dest)
 	 		 	 {
 	 		 	 	count2 = count>>2;
 	 		 	 	count &= 3;
-	 		 	 	longdest = (ulong *) dest;
+	 		 	 	longdest = (uint32_t *) dest;
 	 		 	 	
 	 		 	 	while (count2--)
 	 		 	 	 	*(longdest++) = longval;
@@ -135,8 +135,8 @@ uchar *gr_rsd8_unpack(uchar *src, uchar *dest)
 	 		 	 {
 	 		 	 	count2 = count>>2;
 	 		 	 	count &= 3;
-	 		 	 	longdest = (ulong *) dest;
-	 		 	 	longsrc = (ulong *) src;
+	 		 	 	longdest = (uint32_t *) dest;
+	 		 	 	longsrc = (uint32_t *) src;
 	 		 	 	
 	 		 	 	while (count2--)
 	 		 	 	 	*(longdest++) = *(longsrc++);
@@ -155,11 +155,11 @@ uchar *gr_rsd8_unpack(uchar *src, uchar *dest)
 	 		 	
 	 		 	if (count>=kMinLongLoop)		// if at least kMinLongLoop bytes, do long word stuff
 	 		 	 {
-	 		 	 	longval = val + (((ulong) val)<<8);
+	 		 	 	longval = val + (((uint32_t) val)<<8);
 	 		 	 	longval += longval<<16;
 	 		 	 	count2 = count>>2;
 	 		 	 	count &= 3;
-	 		 	 	longdest = (ulong *) dest;
+	 		 	 	longdest = (uint32_t *) dest;
 	 		 	 	
 	 		 	 	while (count2--)
 	 		 	 	 	*(longdest++) = longval;
